Refactor controllers because we share logic
