+++ rackhd/proxy/handleall.go
@@ -1,6 +1,7 @@
 package proxy
 
 import (
+	"fmt"
 	"net/http"
 	"sync"
 
@@ -17,29 +18,47 @@ type HandleAll struct {
 	wg       *sync.WaitGroup
 }
 
-// WriteResponse is ...
-func (a *HandleAll) WriteResponse(rw http.ResponseWriter, rp runtime.Producer) {
-	// for k, v := range e.headers {
-	// 	for _, val := range v {
-	// 		rw.Header().Add(k, val)
-	// 	}
-	// }
-	// if e.code > 0 {
-	// 	rw.WriteHeader(e.code)
-	// } else {
-	// 	rw.WriteHeader(http.StatusInternalServerError)
-	// }
-	a.HandleAllEndpoints(a.Request, rw)
-	//some other function that returns my json interface message
-
-	if err := rp.Produce(rw, a.Response.Body); err != nil {
-		panic(err)
-	}
-}
-
-// HandleAllMiddleware is ...
+// HandleAllMiddleware handles the route call
 func HandleAllMiddleware(r *http.Request) middleware.Responder {
 	return &HandleAll{
 		Request: r,
+	}
+}
+
+// WriteResponse is ...
+func (a *HandleAll) WriteResponse(rw http.ResponseWriter, rp runtime.Producer) {
+	addrMap, err := GetStoredAddresses("")
+	resp := Response{}
+	if len(addrMap) == 0 {
+
+		resp.Body = []byte("{No endpoints under management.}")
+		resp.StatusCode = http.StatusOK
+
+	} else if err != nil {
+
+		resp.StatusCode = http.StatusInternalServerError
+		resp.Body = []byte(fmt.Sprintf("{%s}", err))
+
+	} else {
+
+		oneResp := GetResponses(a.Request, addrMap)
+		RespCheck(oneResp, &resp)
+
+	}
+
+	for k, v := range resp.Header {
+		for _, val := range v {
+			rw.Header().Add(k, val)
+		}
+	}
+	// rw.Header().Set("Content-Type", "application/json; charset=utf-8")
+	if resp.StatusCode > 0 {
+		rw.WriteHeader(resp.StatusCode)
+	} else {
+		rw.WriteHeader(http.StatusInternalServerError)
+	}
+
+	if err := rp.Produce(rw, resp.Body); err != nil {
+		panic(err)
 	}
 }
